Tree locations from Barro Colorado Island

Tree locations from Barro Colorado Island

bei

Format

A dataframe, from the spatstat.data package.

This is an is an object of class "ppp" representing the point pattern of tree locations. See ppp.object for details of the format.

The dataset bei gives the positions of 3605 trees of the species Beilschmiedia pendula (Lauraceae) in a 1000 by 500 metre rectangular sampling region in the tropical rainforest of Barro Colorado Island. All spatial coordinates are given in metres. These data are part of a much larger dataset containing the positions of hundreds of thousands of trees belong to thousands of species; see Hubbell and Foster (1983), Condit, Hubbell and Foster (1996) and Condit (1998).

The present data were analysed by Moller and Waagepetersen (2007)

Condit, R. (1998) Tropical Forest Census Plots. Springer-Verlag, Berlin and R.G. Landes Company, Georgetown, Texas.

Condit, R., Hubbell, S.P and Foster, R.B. (1996) Changes in tree species abundance in a neotropical forest: impact of climate change. Journal of Tropical Ecology, 12, 231--256.

Hubbell, S.P and Foster, R.B. (1983) Diversity of canopy trees in a neotropical forest and implications for conservation. In: Tropical Rain Forest: Ecology and Management (eds. S.L. Sutton, T.C. Whitmore and A.C. Chadwick), Blackwell Scientific Publications, Oxford, 25--41.

Moller, J. and Waagepetersen, R.P. (2007) Modern spatial point process modelling and inference (with discussion). Scandinavian Journal of Statistics 34, 643--711.

Source

https://github.com/spatstat/spatstat.data