Estimated marginal means with the emmeans package with sdmTMB

Methods for using the emmeans package with sdmTMB. The emmeans package computes estimated marginal means for the fixed effects.

References

https://aosmith.rbind.io/2019/03/25/getting-started-with-emmeans/

Examples

mesh <- make_mesh(pcod_2011, c("X", "Y"), cutoff = 20)
fit <- sdmTMB(
  present ~ as.factor(year),
  data = pcod_2011, mesh = mesh,
  family = binomial()
)
fit
#> Spatial model fit by ML ['sdmTMB']
#> Formula: present ~ as.factor(year)
#> Mesh: mesh (isotropic covariance)
#> Data: pcod_2011
#> Family: binomial(link = 'logit')
#>  
#>                     coef.est coef.se
#> (Intercept)            -0.61    0.55
#> as.factor(year)2013     0.99    0.22
#> as.factor(year)2015     0.75    0.22
#> as.factor(year)2017     0.01    0.22
#> 
#> Matérn range: 48.45
#> Spatial SD: 1.84
#> ML criterion at convergence: 564.495
#> 
#> See ?tidy.sdmTMB to extract these values as a data frame.
emmeans::emmeans(fit, ~ year)
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
emmeans::emmeans(fit, pairwise ~ year)
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
emmeans::emmeans(fit, pairwise ~ year, type = "response")
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
emmeans::emmeans(fit, pairwise ~ year, adjust = "none")
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"

e <- emmeans::emmeans(fit, ~ year)
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
plot(e)
#> Error in eval(expr, envir, enclos): object 'e' not found

e <- emmeans::emmeans(fit, pairwise ~ year)
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
confint(e)
#> Error in eval(expr, envir, enclos): object 'e' not found
summary(e, infer = TRUE)
#> Error in eval(expr, envir, enclos): object 'e' not found
as.data.frame(e)
#> Error in eval(expr, envir, enclos): object 'e' not found

# interaction of factor with continuous predictor:
fit2 <- sdmTMB(
  present ~ depth_scaled * as.factor(year),
  data = pcod_2011, mesh = mesh,
  family = binomial()
)
fit2
#> Spatial model fit by ML ['sdmTMB']
#> Formula: present ~ depth_scaled * as.factor(year)
#> Mesh: mesh (isotropic covariance)
#> Data: pcod_2011
#> Family: binomial(link = 'logit')
#>  
#>                                  coef.est coef.se
#> (Intercept)                         -0.75    0.48
#> depth_scaled                        -0.98    0.25
#> as.factor(year)2013                  1.03    0.23
#> as.factor(year)2015                  0.79    0.23
#> as.factor(year)2017                  0.01    0.23
#> depth_scaled:as.factor(year)2013    -0.16    0.26
#> depth_scaled:as.factor(year)2015     0.03    0.26
#> depth_scaled:as.factor(year)2017    -0.01    0.26
#> 
#> Matérn range: 33.38
#> Spatial SD: 2.19
#> ML criterion at convergence: 546.074
#> 
#> See ?tidy.sdmTMB to extract these values as a data frame.
# slopes for each level:
emmeans::emtrends(fit2, ~ year, var = "depth_scaled")
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
# test difference in slopes:
emmeans::emtrends(fit2, pairwise ~ year, var = "depth_scaled")
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"
emmeans::emmip(fit2, year ~ depth_scaled,
  at = list(depth_scaled = seq(-2.5, 2.5, length.out = 50)), CIs = TRUE)
#> Error in validObject(.Object): invalid class “emmGrid” object: invalid object for slot "bhat" in class "emmGrid": got class "AsIs", should be or extend class "numeric"