Cross validation with sdmTMB models

Facilitates cross validation with sdmTMB models. Returns the log likelihood of left-out data, which is similar in spirit to the ELPD (expected log pointwise predictive density). The function has an option for leave-future-out cross validation. By default, the function creates folds randomly but folds can be manually assigned via the fold_ids argument.

Usage

sdmTMB_cv(
  formula,
  data,
  mesh_args,
  mesh = NULL,
  time = NULL,
  k_folds = 8,
  fold_ids = NULL,
  lfo = FALSE,
  lfo_forecast = 1,
  lfo_validations = 5,
  parallel = TRUE,
  use_initial_fit = FALSE,
  spde = deprecated(),
  ...
)

Arguments

formula

Model formula.

data

A data frame.

mesh_args

Arguments for make_mesh(). If supplied, the mesh will be reconstructed for each fold.

mesh

Output from make_mesh(). If supplied, the mesh will be constant across folds.

time

The name of the time column. Leave as NULL if this is only spatial data.

k_folds

Number of folds.

fold_ids

Optional vector containing user fold IDs. Can also be a single string, e.g. "fold_id" representing the name of the variable in data. Ignored if lfo is TRUE

lfo

Whether to implement leave-future-out (LFO) cross validation where data are used to predict future folds. time argument in sdmTMB() must be specified. See Details section below.

lfo_forecast

If lfo = TRUE, number of time steps to forecast. Time steps 1, ..., T are used to predict T + lfo_forecast and the last forecasted time step is used for validation. See Details section below.

lfo_validations

If lfo = TRUE, number of times to step through the LFOCV process. Defaults to 5. See Details section below.

parallel

If TRUE and a future::plan() is supplied, will be run in parallel.

use_initial_fit

Fit the first fold and use those parameter values as starting values for subsequent folds? Can be faster with many folds.

spde

Depreciated. Use mesh instead.

...

All other arguments required to run sdmTMB() model with the exception of weights, which are used to define the folds.

Value

A list:

data: Original data plus columns for fold ID, CV predicted value, and CV log likelihood.

models: A list of models; one per fold.

fold_loglik: Sum of left-out log likelihoods per fold.

sum_loglik: Sum of fold_loglik across all left-out data.

pdHess: Logical vector: Hessian was invertible each fold?

converged: Logical: all pdHess TRUE?

max_gradients: Max gradient per fold.

Prior to sdmTMB version '0.3.0.9002', elpd was incorrectly returned as the log average likelihood, which is another metric you could compare models with, but not ELPD. For maximum likelihood, ELPD is equivalent in spirit to the sum of the log likelihoods.

Details

Parallel processing

Parallel processing can be used by setting a future::plan().

For example:

library(future)
plan(multisession)
# now use sdmTMB_cv() ...

Leave-future-out cross validation (LFOCV)

An example of LFOCV with 9 time steps, lfo_forecast = 1, and lfo_validations = 2:

Fit data to time steps 1 to 7, predict and validate step 8.

Fit data to time steps 1 to 8, predict and validate step 9.

An example of LFOCV with 9 time steps, lfo_forecast = 2, and lfo_validations = 3:

Fit data to time steps 1 to 5, predict and validate step 7.

Fit data to time steps 1 to 6, predict and validate step 8.

Fit data to time steps 1 to 7, predict and validate step 9.

See example below.

Examples

mesh <- make_mesh(pcod, c("X", "Y"), cutoff = 25)

# Set parallel processing first if desired with the future package.
# See the Details section above.

m_cv <- sdmTMB_cv(
  density ~ 0 + depth_scaled + depth_scaled2,
  data = pcod, mesh = mesh,
  family = tweedie(link = "log"), k_folds = 2
)
#> Running fits with `future.apply()`.
#> Set a parallel `future::plan()` to use parallel processing.

m_cv$fold_loglik
#> [1] -3387.398 -3329.670
m_cv$sum_loglik
#> [1] -6717.068

head(m_cv$data)
#> # A tibble: 6 × 15
#>    year     X     Y depth density present   lat   lon depth_mean depth_sd
#>   <int> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl>      <dbl>    <dbl>
#> 1  2003  446. 5793.   201   113.        1  52.3 -130.       5.16    0.445
#> 2  2003  446. 5800.   212    41.7       1  52.3 -130.       5.16    0.445
#> 3  2003  449. 5802.   220     0         0  52.4 -130.       5.16    0.445
#> 4  2003  437. 5802.   197    15.7       1  52.4 -130.       5.16    0.445
#> 5  2003  421. 5771.   256     0         0  52.1 -130.       5.16    0.445
#> 6  2003  418. 5772.   293     0         0  52.1 -130.       5.16    0.445
#> # ℹ 5 more variables: depth_scaled <dbl>, depth_scaled2 <dbl>, cv_fold <int>,
#> #   cv_predicted <dbl>, cv_loglik <dbl>
m_cv$models[[1]]
#> Spatial model fit by ML ['sdmTMB']
#> Formula: density ~ 0 + depth_scaled + depth_scaled2
#> Family: tweedie(link = 'log')
#>  
#>               coef.est coef.se
#> depth_scaled     -2.06    0.19
#> depth_scaled2    -1.27    0.10
#> 
#> Dispersion parameter: 13.10
#> Tweedie p: 1.60
#> Matérn range: 96.42
#> Spatial SD: 2.78
#> ML criterion at convergence: 3206.009
#> 
#> See ?tidy.sdmTMB to extract these values as a data frame.
m_cv$max_gradients
#> [1] 1.814759e-10 3.560097e-12

# \donttest{
# Create mesh each fold:
m_cv2 <- sdmTMB_cv(
  density ~ 0 + depth_scaled + depth_scaled2,
  data = pcod, mesh_args = list(xy_cols = c("X", "Y"), cutoff = 20),
  family = tweedie(link = "log"), k_folds = 2
)
#> Running fits with `future.apply()`.
#> Set a parallel `future::plan()` to use parallel processing.

# Use fold_ids:
m_cv3 <- sdmTMB_cv(
  density ~ 0 + depth_scaled + depth_scaled2,
  data = pcod, mesh = mesh,
  family = tweedie(link = "log"),
  fold_ids = rep(seq(1, 3), nrow(pcod))[seq(1, nrow(pcod))]
)
#> Running fits with `future.apply()`.
#> Set a parallel `future::plan()` to use parallel processing.

# LFOCV:
m_lfocv <- sdmTMB_cv(
  present ~ s(year, k = 4),
  data = pcod,
  mesh = mesh,
  lfo = TRUE,
  lfo_forecast = 2,
  lfo_validations = 3,
  family = binomial(),
  spatiotemporal = "off",
  time = "year" # must be specified
)
#> Running fits with `future.apply()`.
#> Set a parallel `future::plan()` to use parallel processing.
#> Error in `$<-`(`*tmp*`, "cv_loglik", value = c(-2.0274638698412, -0.0210808278056716, -0.0194247223135517, -0.654376030893968, -0.186469931076218, -0.160293756625306, -0.184425509056786, -0.288904966608913, -0.462347679536246, -0.136526965713471, -0.0388369888283875, -0.0392128267920756, -0.145187751135301, -0.129559949589776, -0.0498161725622799, -0.127833360684468, -0.0641453494917986, -0.386320913071439, -0.0633285325297866, -0.21037546742889, -0.139624378782622, -0.128146540558065, -0.0886726280424021, -0.16655034194462, -0.253857240287997, -0.470011319686046, -0.131472487936779, -0.172277338134131, -0.122708953140219, -0.180498466788055, -0.135687252821284, -0.134882026970112, -0.131188823887721, -0.204060650867341, -0.138614269707545, -0.127031547322161, -0.139804030886377, -0.138822423676155, -1.50872989683002, -1.52380867497136, -0.205635380888444, -0.277660242615028, -1.09825682547219, -1.14853015468327, -1.50499407535189, -0.358962778843018, -0.478815874450531, -0.437010047269326, -1.87259598537266, -1.1818025724223, -0.341869437425547, -1.16125379432638, -0.4932404883542, -1.07790254765484, -0.36529210154678, -0.295138766509482, -1.72737101192528, -0.27018851241128, -0.354322898252988, -0.386892967217605, -0.258988041328885, -0.312257500535512, -0.665680397499802, -0.34887475207107, -0.876172087441506, -1.23011088701971, -0.456247749676015, -0.424851556944452, -0.354707189218617, -1.2990476892007, -0.424566494048891, -0.886858084090395, -0.512968583912383, -0.141681128257987, -0.486526258598931, -0.230305861914923, -0.332427564701327, -0.553802982371621, -0.177220591211202, -0.343693593090235, -0.18915987618043, -1.22137545981112, -0.274258611812484, -0.230189162567584, -0.449038606104769, -1.36907982804339, -1.0746486425962, -0.261200300084761, -0.371076566514457, -0.53167232958395, -0.918466597915797, -0.296744243885217, -0.49520495456261, -0.857721158064506, -0.0701763516468404, -0.38265491442821, -0.450003753901594, -0.391178248797196, -0.327652813034148, -0.313881317869912, -0.780647039672428, -0.45014573090338, -0.311441198558008, -0.326017344086163, -0.777865312068378, -0.44794060560177, -0.460198795262343, -0.559924971630302, -0.3683500072553, -0.0692528018410657, -0.566774229363479, -0.835071100637967, -0.721538957962901, -0.747416311593857, -0.0535264686477566, -0.850212647639991, -0.680023954895367, -0.620949059658101, -0.73975352181102, -0.3316083775742, -0.216097376647843, -0.148080510972825, -0.383943546522784, -0.234668807665461, -0.735203423041172, -0.433442369075567, -0.54058331575882, -0.1473508607331, -0.213889903688263, -0.676905607209966, -0.691760909273284, -1.11668839159875, -0.262795492579203, -0.302641903935928, -0.911313420755587, -1.50160071884201, -0.824555841179385, -0.360986945421742, -0.344904881025642, -0.335754705803353, -1.14379881980157, -1.84514402421752, -0.397962134537378, -0.728666192043936, -0.589921103513298, -0.0835598735160129, -0.493327153424544, -0.699724344319701, -0.510799708187421, -1.02736569640532, -0.395629318507735, -1.01721127256484, -0.0975433587050968, -0.248897695094161, -0.97009481316387, -0.516529078174921, -0.71134585312319, -0.723976489716568, -0.107370580020215, -0.76008496187729, -0.611290290074344, -0.643761129687424, -0.499649268633624, -0.401892330537275, -0.460257838985039, -0.417931886928078, -0.788983085088609, -1.16442939133343, -0.830669545303266, -1.04697954174828, -0.517075799367788, -0.505350332385572, -0.332575249607059, -1.09475167933303, -0.387666807683184, -0.546707139382991, -0.366461836912147, -0.312766012395042, -0.501376684460983, -0.393252109627384, -0.313035569972949, -0.468349588872175, -1.01376462552025, -0.818217990233762, -0.756684884333542, -0.555688402639264, -0.864266089501943, -1.29296976808495, -0.865995908560839, -0.814848388006587, -0.36471002812601, -0.321210869793372, -1.30410479642472, -0.512932889559465, -0.906825273374042, -0.660518284925252, -1.47443645562368, -1.49725408712676, -0.0141539872837352, -0.335526047851415, -0.276412700743214, -0.342369762406832, -0.343750611340237, -0.650714046380353, -0.963982373199708, -0.446576504410196, -0.689902363129458, -0.964564225789477, -1.46370034581335, -0.233401537715219, -1.32382343920978, -1.29250336988336, -0.285082651090611, -0.202723561606865, -0.27040258964044, -0.348971378792211, -0.836355765337622, -0.737284371041323, -0.219330730654788, -0.284173025965456, -0.581684030940538, -0.725169502804153, -0.295454440879332, -0.312911458578676, -0.316589243093654, -0.712333498916237, -0.156608314326712, -0.189250199062021, -0.716255730882769, -0.675115831969702, -0.431635732579685, -0.0665087613260607, -0.685502320372803, -0.747839750430029, -0.817734698765465, -0.0144469388226788, -0.691027305993852, -1.05426974446138, -0.0408676097738973, -0.606421521846879, -0.540966788470079, -0.336362317330227, -0.511901660118391, -0.625116419748007, -0.540278524135671, -0.0474486227113093, -0.0606022759097217, -0.826012916172512, -0.856280196270162, -0.746897422703958, -0.540882190236552, -1.13413768205824, -0.789083516522651, -0.435897721714474, -1.08701579839253, -0.335664328833014, -0.832564572091667, -0.495494370241464, -1.9170945568541, -0.377859362543383, -1.5508493858332, -0.862730573855048, -2.08669155801365, -0.374133130899009, -0.898958313396316, -0.576946127242313, -2.10511841164606, -0.360383179814362, -1.66064361454581, -0.132849768886129, -0.568413584896698, -0.522514402205579, -1.96521601110802, -0.644819844102502, -0.502649779851365, -0.532231472823957, -0.628255756335084, -0.306137060112728, -0.491637981298158, -0.615129839284459, -0.38669423594944, -0.424190553426748, -0.507815260946832, -0.332897551351954, -0.341444109848213, -0.7117711384348, -0.581740976827283, -0.348945312906116, -0.592724191301044, -0.610287679004437, -0.656339722694843, -0.578395157363402, -0.357461925066316, -0.917813824716994, -0.731536078654594, -1.88240658876388, -0.471518317187442, -0.516060815191859, -0.542364763827791, -0.783319857637537, -0.506439062087233, -0.369417487617788, -0.457584988468154, -0.157769928498341, -0.0135434346004197, -0.600617158375367, -0.924931892086406, -0.0436713824389344, -0.0389143536966907, -0.0318738793343328, -0.451297784881338, -0.944222014219965, -0.675621446783388, -0.457265921320827, -0.737931389990255, -0.0499518225538516, -2.67935557630523, -0.203072559936314, -1.00072956692424, -0.110070837271499, -0.554407184132714, -0.198955953366155, -0.492987862502737, -0.0463980963628494, -1.66734980435736, -0.671985904570845, -0.912000714457258, -0.993415480310454, -0.790999397636239, -0.774838306620334, -0.816476930598968, -0.0741785405104775, -0.078035456023504, -0.975978457343813, -0.705719556780112, -1.00856476142771, -1.0155970800818, -0.641580347821688, -0.837360640155338, -0.852638608666616, -0.778124046692314, -0.730685049428009, -1.65071171813035, -2.11027218849726, -1.6399548737385, -1.66309066507804, -1.40824347288636, -0.249006980596012, -2.06611262720599, -1.73635799066162, -1.09538880405602, -0.249681066217264, -1.49195187144254, -1.4523444803715, -1.21210936058381, -1.26572106987623, -1.1710641724178, -0.25655330398962, -1.26987149945506, -0.586010207061011, -0.454855739798436, -1.4413398437339, -0.562524125217931, -0.325519080817786, -0.919452787652573, -0.240279253635028, -0.803753091290988, -0.533814451896788, -0.14282272795018, -0.489832583719352, -0.850646165871193, -0.69891681094181, -0.481902664978322, -1.39079907096024, -0.125522866214885, -1.04909794047079, -0.650110874317407, -0.604151111624189, -0.628779822815416, -2.18773725449797, -1.08987942727408, -1.01753009439425, -0.129458859607246, -0.11989478073364, -0.135470428072121, -0.0209980328892032, -0.604860010888142, -0.129716206303035, -0.0407686490363615, -0.0308763715826068, -1.14796172258864, -0.102132671859113, -0.329152877610033, -0.779536130238659, -1.13331940746418, -0.28529315191057, -0.744840824752247, -0.975842918403461, -0.4156969077886, -0.352381675395628, -1.1844630017391, -0.270787122409415, -0.333668212210189, -1.2759475334623, -0.301699746521703, -0.318211332907602, -0.753289305425584, -1.17271190137423, -0.350097566188444, -0.381317226234241, -1.07778402853341, -1.13025353716135, -0.271632568082704, -0.579295175473077, -0.902483491053678, -0.778179998271069, -0.343315391836236, -0.577945642268361, -0.422562535789778, -0.362236088881227, -0.935668094000938, -0.83040645849597, -0.11710286610391, -0.115320292034891, -0.0489711159901443, -0.0523829834626696, -0.124674548797524, -0.116755633346063, -2.2825117889231, -0.0849539545450569, -0.146723458931173, -1.41480926003001, -0.123875342152844, -0.116835812081561, -0.113124518748944, -0.112279836691065, -0.0885485430626347, -0.202275671062113, -2.04783596127116, -0.115670663278647, -0.204502171117616, -0.119467283648928, -2.06790536165919, -0.120774956588008, -1.18597636829049, -0.239972312642675, -0.136676559380583, -2.14167296163358, -2.14758190329644, -1.35504851295905, -0.322854423872056, -0.336310082200113, -0.427220475754099, -0.254649706036898, -0.908619975270837, -0.459300679058846, -0.265115320933965, -0.410547532060637, -0.418743292278787, -1.10462240498831, -0.744609789680051, -0.337751039632343, -0.200435854725376, -0.989347286600728, -0.138925963581711, -0.270963277156538, -0.243034066555405, -0.167811040029002, -0.326727837559904, -1.75475832830613, -1.33690628208184, -0.269499211134264, -1.11082532626229, -0.217376555280287, -1.03946991087821, -0.324350682818777, -0.911568229768495, -0.970320415583611, -0.164840076639681, -0.174846146132295, -0.92106052453076, -0.27167048042948, -0.71787894454326, -0.980224477684235, -0.160116004899793, -0.148532787161346, -0.200669683293505, -0.599271498201435, -0.721058536750094, -0.261664778005634, -0.144220302657004, -1.33021004609038, -0.232226412802338, -0.288407077725696, -0.285908782397608, -0.293597327321732, -0.306801783278267, -1.05721456284724, -0.868551296252393, -0.970652476332792, -0.352139914707349, -0.652041936769158, -0.828245689168938, -0.743150782626572, -0.528187046800305, -0.517837283601607, -0.6201970865968, -0.789778283344579, -0.86050744171667, -0.388280922883298, -0.485064083166266, -1.92511680879899, -0.308181077207629, -1.37363681577266, -1.05883043921159, -0.232930782073363, -0.514542973241788, -0.27180523330659, -0.498545110265501, -0.248346792439083, -0.288463549000535, -0.27590788872586, -1.45746350970091, -1.10005481096061, -1.03551495118552, -1.10302056292808, -0.877400231765816, -0.981742709331186, -1.02776660109409, -1.01380258619181, -1.08965303457539, -0.970585305378181, -1.07310979852936, -0.419233906435208, -0.808163876809978, -0.383128195582597, -0.428576031237846, -1.3028800455894, -1.32855137399403, -0.17640587447343, -0.145217417266502, -1.7614565089253, -1.25190212532119, -1.51658420026439, -0.548691760551511, -0.550281077957587, -0.989357905930122, -1.84630640777073, -0.225247701932778, -1.25512013631297, -0.143215180031597, -2.20768191713793, -0.116512677089456, -2.05806997965456, -0.139515776426701, -0.138743651690524, -0.830150912677947, -0.317149778463952, -0.365488003793483, -1.03023213972114, -0.922680078323576, -0.559712640721047, -1.23699111643817, -0.685698204057195, -0.362013298550591, -0.743624047875834, -0.214302035320125, -0.247474060663092, -0.418756580190465, -0.342140566899006, -1.93008336956646, -0.372002201727236, -0.325492429604518, -0.741427004809396, -0.170718809861062, -0.189023769143424, -0.263221459150663, -0.299928695116783, -0.197043387743873, -0.286954687313048, -0.559297720577667, -0.566203008611461, -0.473103377190423, -0.419670989664924, -1.20815390816942, -0.501037179427192, -0.742751195736458, -0.743859527410683, -1.02854230322432, -0.618467721556471, -0.444097234851062, -0.67155531450307, -0.604757252671459, -0.536309179572646, -0.554840255459718, -0.41389237000509, -0.507408675595407, -0.780298115019866, -0.571043515645285, -0.356825252911287, -0.596484694279998, -0.557096103444491, -0.982835498364815, -0.612932060350576, -0.223152134969784, -0.425985777764288, -1.12221574077645, -0.993084387096128, -0.604442738417026, -1.04358906697656, -0.825858252632168, -2.54643078533167, -0.206596141169447, -0.189497912654059, -0.216400413806503, -0.221188105808257, -1.54229559647396, -1.48893928450997, -0.171501638149884, -0.200361468798412, -0.328133313488568, -0.182868276287604, -0.156033864255072, -0.514158046857974, -0.303071783513453, -1.69595942302343, -0.851453398524, -0.139477525516361, -0.667277038664267, -0.553441469253371, -0.336652816540473, -0.574926177207401, -0.386449937310663, -0.292138249238765, -0.3927471803031, -0.497816341653716, -0.552613370266833, -0.962963590302803, -0.476851475505934, -0.837700475599651, -0.417169637123324, -1.23018990362617, -1.37589901253538, -0.0940274671407222, -0.0456964339211715, -0.0535077135115667, -0.0391898473210934, -0.1514768693182, -1.80361284324934, -1.73356428670035, -0.501909031065429, -0.50627165199702, -1.13667785336582, -0.362782810874281, -2.36532259436166, -0.108994543583352, -0.0988655080122361, -0.0956122099060525, -0.326163660760496, -0.101806624034271, -0.120311660076979, -0.128066259871987, -0.104306404013059, -0.103591723812981, -0.313323950450516, -1.33649919464982, -0.310433261903677, -0.276134667008686, -0.279867292675008, -1.22740947669172, -1.31000876918251, -1.1379633023295, -1.11752363137367, -0.30781627967569, -1.02985879625453, -1.22050706892728, -1.72467190393082, -1.41982403640599, -0.236195649777054, -0.245341178219781, -0.11951109733088, -1.75801495263564, -0.32393799982912, -0.323724550556649, -1.28111146240644, -0.332074897089926, -1.29239133775946, -0.357188496197386, -1.23027468789643, -0.278013079771229, -1.67966009945862, -0.110443586455887, -0.0703105051209133, -0.090300711336103, -2.80042996930535, -1.54318722815439, -0.0313771248526576, -0.0403993663339521, -0.100449763536378, -2.26208469111603, -0.102324255654816, -0.10529147520907, -0.104178714988317, -0.101728903061699, -0.154310477954189, -1.85876784148702, -2.26687720851407, -0.069022869057042, -0.0538307675358395, -2.25355040207171, -0.0377448432251469, -0.0395479301581852, -0.697338896725554, -0.603148431716725, -0.0109903448951189, -0.482177984602682, -0.97673068052437, -0.621822388961152, -0.549040448188731, -1.01432474738096, -0.562835220778796, -0.623873971719695, -0.374120406204188, -0.380394150631075, -0.443922179867004, -0.910826764348813, -0.576898056655923, -0.815656650441114, -0.793596767421455, -2.40545547942677, -0.0912486722292311, -0.0905846583850688, -2.38874910307521, -2.36152724358741, -0.100068175302495, -0.308584286555033, -0.471069363098602, -0.0876852655441961, -0.0861021847968718, -0.0913464549082339, -1.80960120468395, -0.129932841063785, -0.147297767066998, -1.79387372624424, -0.14637417340965, -0.165573003718461, -0.0623035640733621, -0.0329757933717271, -0.0271192459991783, -0.0505113431358058, -0.0380094086488228, -0.020167928018619, -0.0167602829992781, -0.356724116170909, -0.373199127151924, -0.407332408534983, -0.361774266521754, -0.556993642580298, -0.43487080637381, -0.446188382154082, -0.511138519079513, -0.906146997090461, -0.443534332805165, -0.393725764492488, -0.500345866347122, -1.27382982736299, -0.437010759397226, -0.0944485263892396, -2.51663143016479, -0.230027126293201, -0.073267312105689, -0.995633877620587, -0.0923493860247704, -0.0934798868507366, -0.0934118668176932, -0.0925547760580299, -0.0916552229442405, -0.0957095678123505, -0.0935849062219925, -0.450035251452124, -1.80501920763701, -0.239009223278876, -0.344998817162346, -0.274865072771785, -0.0493070132788222, -0.151280727476776, -0.0810720486431693, -0.101688449303843, -0.186233854087966, -0.106600962033148, -0.117813719727174, -0.109157764099689, -0.107934210161482, -0.101174850400687, -0.10508102678401, -0.135221980224842, -0.143217744130362, -0.192444953849056, -1.87511578215268, -0.267802802544531, -1.47180844663537, -0.172165255068582, -0.236824199489134, -0.200075615155026, -0.181697795023502, -0.0304845629088519, -0.0233169988615755, -0.0479280354223358, -0.0439352676463446, -0.238194251228783, -0.216904019310958, -1.63350641863686, -0.17981618955834, -1.14232007750696, -0.476062503464916, -1.09032307298683, -0.540602818256699, -0.65391983242094, -0.00926931739114769, -0.045504624940428, -0.0372995445784225, -0.0509434371314185, -0.0263742827718149, -0.0531799145197967, -0.278631740439067, -0.397978512858817, -0.720496057708444, -0.58302594459642, -0.758229139752531, -0.565986902335671, -0.405007393763157, -0.496732643400421, -1.06112840714171, -0.706202512741834, -0.273519787373796, -0.196713728648775, -0.209097692992554, -1.67034512777412, -1.59903356440752, -0.689358415918732, -0.888381571512059, -2.13373705177409, -2.05643477864806, -0.149963759861065, -0.224987268191357, -0.309358822897185, -1.57841235519802, -0.178432264828273, -1.78460399416844, -0.185262783651308, -1.0799612351489, -0.388527962206845, -0.27834332899076, -0.304968433532094, -0.710376983584886, -0.718968974759802, -0.537875755963597, -0.565292241387618, -0.895985132520048, -1.41222510698942, -0.446002144156147, -0.810259733292157, -0.341924524755536, -0.686004623302126, -0.665221064563507, -0.968473914868686, -1.3127254732636, -1.51397525214142, -1.75058331355966, -0.414964447257594, -0.358537204006681, -0.776156329884846, -0.444117191638889, -0.798501385926786, -0.702937445329206, -0.79692148086427, -0.402409114206708, -0.900862027774242, -0.588340435524515, -0.78953152160302, -0.508357640806621, -0.507421857510109, -0.366586059085267, -0.354379313715185, -0.280039215862708, -0.587328607186781, -0.257565393990878, -0.249048879663365, -0.238737217222068, -0.246653676270559, -0.593037208022544, -0.585793097072802, -0.669442454959736, -0.510013828292524, -0.571044761514831, -0.487083279877733, -0.422848531785623, -0.988871679561755, -0.387427480753458, -1.13911124952177, -0.481811800724011, -0.405714207684683, -1.53684322966418, -0.222785688293699, -2.32547243865546, -2.0222557480262, -0.460346496810142, -0.318213861558017, -0.108554140897031, -0.128920710601269, -0.282646957874803, -0.0941675062857028, -1.01919154054278, -0.575194718788701, -0.552452820074277, -0.684009589479364, -0.871580643683957, -0.79233257214649, -0.86872275861549, -1.48241936614852, -0.287400991232543, -1.57555573015299, -0.791270424755292, -0.325111682820844, -0.979980861911475, -0.282836006362321, -1.52239015075616, -1.789805776504, -0.147731635567087, -0.187515314929224, -0.654295464518517, -0.710935180640025, -0.679670601519803, -0.688350553710183, -0.681994881784832, -0.617443514108911, -0.410311010613155, -1.04180125429925, -0.668142952203371, -1.15480714774549, -0.960364606888985, -0.329386186582209, -1.46473936374331, -1.60331313501881, -1.2191688498421, -0.812046447856005, -0.397623336669115, -0.159088568370114, -0.248710433729854, -0.24827599958161, -0.443334528136256, -0.531672415413949, -0.127974438404019, -0.124155025094337, -0.269306435887318, -0.256356155355864, -0.204569337974386, -0.251823291173575, -0.253645345324616, -0.402634413560456, -1.09280865663113, -0.439217955087749, -1.1502049928795, -0.387835912149175, -0.823085087026568, -0.480560723947704, -0.883314546575448, -1.18016446902245, -1.00833223569532)): Assigned data `cv_loglik[tmb_data$weights_i == 1]` must be compatible
#> with existing data.
#> ✖ Existing data has 240 rows.
#> ✖ Assigned data has 969 rows.
#> ℹ Only vectors of size 1 are recycled.
#> Caused by error in `vectbl_recycle_rhs_rows()`:
#> ! Can't recycle input of size 969 to size 240.

# See how the LFOCV folds were assigned:
example_data <- m_lfocv$models[[1]]$data
#> Error in eval(expr, envir, enclos): object 'm_lfocv' not found
table(example_data$cv_fold, example_data$year)
#> Error in eval(expr, envir, enclos): object 'example_data' not found
# }