Save log likelihoods of k-fold cross-validation for sdmTMB models

Save log likelihoods of k-fold cross-validation for sdmTMB models

sdmTMB_cv(formula, data, spde, time = NULL, k_folds = 10, fold_ids = NULL, ...)

Arguments

formula
Model formula.

data
A data frame.

spde
Output from make_spde().

time
The name of the time column. Leave as NULL if this is only spatial data.

k_folds
Number of folds.

fold_ids
Optional input name of column containing user fold ids.

...
All other arguments required to run sdmTMB() model with the exception of weights, which are used to define the folds.

Examples

d <- subset(pcod, year >= 2011) # subset for example speed
spde <- make_spde(d$X, d$Y, n_knots = 30)

# library(future) # for parallel processing
# plan(multisession) # for parallel processing
m_cv <- sdmTMB_cv(
  density ~ 0 + depth_scaled + depth_scaled2 + as.factor(year),
  data = d, spde = spde,
  family = tweedie(link = "log"), time = "year", k_folds = 3
)
m_cv$fold_loglik

#> [1] -1967.241 -2333.340 -2572.307

m_cv$sum_loglik

#> [1] -6872.889

head(m_cv$data)

#> # A tibble: 6 x 15
#>    year     X     Y depth density present   lat   lon depth_mean depth_sd
#>   <int> <dbl> <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl>      <dbl>    <dbl>
#> 1  2011  435. 5718.   241    245.       1  51.6 -130.       5.16    0.445
#> 2  2011  435. 5718.   241    245.       1  51.6 -130.       5.16    0.445
#> 3  2011  487. 5719.    52      0        0  51.6 -129.       5.16    0.445
#> 4  2011  487. 5719.    52      0        0  51.6 -129.       5.16    0.445
#> 5  2011  490. 5717.    47      0        0  51.6 -129.       5.16    0.445
#> 6  2011  490. 5717.    47      0        0  51.6 -129.       5.16    0.445
#> # … with 5 more variables: depth_scaled <dbl>, depth_scaled2 <dbl>,
#> #   cv_fold <int>, cv_predicted <dbl>, cv_loglik <dbl>

m_cv$models[[1]]

#> Spatiotemporal model fit by ML ['sdmTMB']
#> Formula: formula
#> SPDE: spde
#> Family: tweedie(link = 'log')
#>                     coef.est coef.se
#> depth_scaled           -1.25    0.25
#> depth_scaled2          -0.93    0.16
#> as.factor(year)2011     4.21    0.35
#> as.factor(year)2013     3.86    0.38
#> as.factor(year)2015     3.57    0.36
#> as.factor(year)2017     2.89    0.42
#> 
#> Matern range parameter: 16.65
#> Dispersion parameter: 14.97
#> Spatial SD (sigma_O): 0.70
#> Spatiotemporal SD (sigma_E): 2.19
#> ML criterion at convergence: 1043.687

m_cv$max_gradients

#> [1] 8.520158e-05 4.929657e-06 9.255972e-05